pcd: actually split stage plumbing back out into sample_eval_commitment helper
